backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
pillow==10.1.0
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
ultralytics==8.0.207
numpy==1.26.2
scipy==1.11.4
shapely==2.0.2